payment-gen/requirements.txt
pyyaml>=6.0
jsonschema>=4.17.0
click>=8.1.0
